pyproject.toml
[tool.poetry]
name = "telegramfinances"
version = "0.1.0"
description = ""
authors = ["Kirill Ibragimov <hexchap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
aiogram = "^2.24"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
uvloop = "^0.17.0"
ujson = "^5.6.0"
pydantic = "^1.10.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
